+++ src/ZODB/scripts/repozo.py
@@ -255,9 +255,9 @@ def parseargs(argv):
         if options.output is not None:
             log('--output option is ignored in backup mode')
             options.output = None
-        if options.withverify is not None:
+        if options.withverify:
             log('--with-verify option is ignored in backup mode')
-            options.withverify = None
+            options.withverify = False
         if not options.file:
             usage(1, '--file is required in backup mode')
     elif options.mode == RECOVER:
@@ -287,9 +287,9 @@ def parseargs(argv):
         if options.killold:
             log('--kill-old-on-full option is ignored in verify mode')
             options.killold = False
-        if options.withverify is not None:
+        if options.withverify:
             log('--with-verify option is ignored in verify mode')
-            options.withverify = None
+            options.withverify = False
     return options